Move search child components to separate files
